Fix incorrect printa_sf calls everywhere
Because Soar's custom output formatter uses non-standard specifiers
(%u for uint64_t, %d for int64_t and %f for double), it's easy to get
the type input types incorrect, which leads to undefined behavior. For
example, in #342 Bob found that when printing the integer 0 with %d,
garbage would be output because an extra 4 bytes of memory were read
for the value. This could potentially lead to the program crashing
due to accessing restricted memory.

Fix these everywhere using `static_cast`. Also fix some places where
the specifiers were wrong in different ways (%u instead of %d, and %d
instead of %p in one case).

Fixes #342.

+++ Core/CLI/src/cli_print.cpp
@@ -127,7 +127,7 @@ void print_stack_trace(agent* thisAgent, bool print_states, bool print_operators
 
     if (stateCount > maxStates)
     {
-        thisAgent->outputManager->printa_sf(thisAgent,  "...Stack goes on for another %d states\n", stateCount - maxStates);
+        thisAgent->outputManager->printa_sf(thisAgent,  "...Stack goes on for another %d states\n", static_cast<int64_t>(stateCount - maxStates));
     }
 }
 
@@ -706,7 +706,7 @@ void print_symbol(agent* thisAgent, const char* arg, bool print_filename, bool i
             }
             if (!w)
             {
-                thisAgent->outputManager->printa_sf(thisAgent,  "No wme %u in working memory.", lexeme.int_val);
+                thisAgent->outputManager->printa_sf(thisAgent,  "No wme %d in working memory.", lexeme.int_val);
             }
             break;
